MG-359: searchResultsList -> searchResults[]
