+++ libs/explorers/models/src/index.ts
@@ -2,6 +2,6 @@ export * from './lib/app-error';
 export * from './lib/loading-icon-colors';
 export * from './lib/navigation-link';
 export * from './lib/panel';
-export * from './lib/results-list';
+export * from './lib/search-result';
 export * from './lib/synapse-wiki';
 export * from './lib/terms-of-use-info';
