--- libs/explorers/models/src/lib/results-list.ts
@@ -1,9 +1,0 @@
-export interface SearchResult {
-  name: string;
-  alias: string[];
-  id: string;
-}
-
-export interface SearchResultsList {
-  items?: SearchResult[];
-}
