+++ libs/explorers/testing/src/lib/mocks/index.ts
@@ -2,6 +2,6 @@ export * from './components/hero-component-mock';
 export * from './components/svg-icon-component-mock';
 export * from './components/wiki-component-mock';
 export * from './navigation-links-mocks';
-export * from './results-list-mocks';
+export * from './search-result-mocks';
 export * from './svg-icon-service-stub';
 export * from './synapse-mocks';
