+++ libs/model-ad/home/src/lib/home.component.html
@@ -29,7 +29,7 @@
           [imageAltText]="'mouse icon'"
           [searchPlaceholder]="'Find model by name...'"
           [navigateToResult]="navigateToResult"
-          [getSearchResultsList]="getSearchResultsList"
+          [getSearchResults]="getSearchResults"
           [checkQueryForErrors]="checkQueryForErrors"
         />
         <explorers-home-card
