+++ libs/model-ad/util/src/lib/search.ts
@@ -1,11 +1,11 @@
-import { SearchResultsList } from '@sagebionetworks/explorers/models';
+import { SearchResult } from '@sagebionetworks/explorers/models';
 import { EMPTY, Observable } from 'rxjs';
 
 export const navigateToResult = (id: string): void => {
   alert(`Navigating to result with ID: ${id}`);
 };
 
-export const getSearchResultsList = (query: string): Observable<SearchResultsList> => {
+export const getSearchResults = (query: string): Observable<SearchResult[]> => {
   return EMPTY;
 };
